_requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.3.2
fsspec==2023.5.0
fugashi==1.2.1
gunicorn==20.1.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pycor==0.0.8
pyocr==0.8.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tokenizers==0.13.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchaudio==2.0.2
torchvision==0.15.2+cpu
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
tzdata==2023.3
unidic-lite==1.0.8
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0